% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.choicemod.R
\name{barplot.choicemod}
\alias{barplot.choicemod}
\title{Barplots of the selected variables}
\usage{
\method{barplot}{choicemod}(res, method = "linreg", type = "varsel", ...)
}
\arguments{
\item{res}{an object of class \code{choicemod}.}

\item{method}{one of the methods in \code{res$method} (among \code{"linreg"}, \code{"sir"}, \code{"rf"},
\code{"pcr"}, \code{"plsr"}, \code{"ridge"}).}

\item{type}{this should be \code{"varsel"} (proportion of selection each variable in the reduced model).
or \code{"modsize"} (frequency of the size of each reduced model).}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Barplots with informations about the variables selected in the reduced models estimated by the function \code{choicemod}.
}
\examples{
data("simus")
res <- simus$res1
class(res)
res$method
barplot(res, method="sir", type="varsel", las = 2,  main = "sir", col = rgb(1,1,0,0.2))
barplot(res, method="sir", type="sizemod", las = 2, main = "sir", col = rgb(0,1,1,0.2))

}
\seealso{
\code{\link{choicemod}}, \code{\link{boxplot.choicemod}}
}
