% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.choicemod.R
\name{barplot}
\alias{barplot}
\title{Generic barplot}
\usage{
barplot(res, ...)
}
\arguments{
\item{res}{an object with the values to plot.}

\item{...}{additional arguments.}
}
\description{
\code{barplot} is a generic function for barplot.
}
