% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.choicemod.R
\name{boxplot.choicemod}
\alias{boxplot.choicemod}
\title{Boxplots for models comparison}
\usage{
\method{boxplot}{choicemod}(res, method = NULL, measure = "mse",
  type = "reduced", ...)
}
\arguments{
\item{res}{an object of class \code{choicemod}.}

\item{method}{if NULL, all the methods in \code{res$method} are plotted. Otherwise a subset of these methods (among \code{"linreg"}, \code{"sir"}, \code{"rf"}).}

\item{measure}{the criterion to be plotted. This should be \code{"mse"} (by default) or \code{"r2"}.}

\item{type}{this should be \code{"reduced"} (boxplot for reduced estimated models),  \code{"complete"} (boxplot for complete estimated models) or \code{"both"}. By default,  \code{type="reduced"}.}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Parallel boxplots of the results (mean square errors or R2 coefficients)  obtained for complete models (with all the input variables) and for reduced models (with a selection of relevant variables) estimated by the function \code{choicemod}.
}
\examples{
data("simus")
res <- simus$res1
class(res)
res$method
boxplot(res, type="reduced", col="gold", main="Models with variables selection, N=10")
boxplot(res,type="complete",col=rgb(0,0,1,0.2), main="Models with all the variables, N=10")
boxplot(res,type="both", col = "aquamarine", las = 2, main = "N = 10 replications")
boxplot(res,method=c("linreg","sir"), measure="r2", col=rgb(0,1,0,0.2))

}
\seealso{
\code{\link{choicemod}}, \code{\link{barplot.choicemod}}
}
