% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.choicemod.R
\name{boxplot}
\alias{boxplot}
\title{Generic parallel boxplots}
\usage{
boxplot(res, ...)
}
\arguments{
\item{res}{an object with the values to plot.}

\item{...}{additional arguments.}
}
\description{
\code{boxplot} is a generic function for parallel boxplots.
}
