% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choicemod.R
\name{choicemod}
\alias{choicemod}
\title{Choice of the regression model}
\usage{
choicemod(X, Y, method = c("linreg", "sir", "rf"), N = 20,
  prop_train = 0.8, nperm = 50, cutoff = TRUE, nbsel = NULL)
}
\arguments{
\item{X}{the input numerical matrix of dimension nxp.}

\item{Y}{the output numerical vector of size n.}

\item{method}{a vector with the methods.
chosen among \code{"linreg"}, \code{"sir"}, \code{"rf"}, \code{"pcr"},
\code{"plsr"}, \code{"ridge"}.}

\item{N}{the number of replications i.e. the number of ramdom leaning/test
samples used to evaluate MSE of the chosen model.}

\item{prop_train}{a value between 0 and 1 with the proportion of
observations in the \code{N} training samples.}

\item{nperm}{the number of repetitions of random perturbations of the
predictors used to calculate the importance of the variables (VI values).}

\item{cutoff}{if TRUE the variables are selected automatically and the number
of selected variables is unknown. If \code{cutoff=FALSE} the \code{nbsel}
 best variables are selected.}

\item{nbsel}{the number of selected variables. Active only if
\code{cutoff=FALSE}.}
}
\value{
An object with S3 class "choicemod" and with the following components:
\item{mse}{ a matrix of dimension \code{N} times \code{length(methods)}.
Each colum  gives the  \code{N} values of MSE calculated on the test samples.
Here only the selected variables are considered in the models estimated on
the training samples.}

\item{mse_all}{a matrix of dimension \code{N} times \code{length(methods)}.
Each colum  gives the  \code{N} values of MSE calculated on the test samples.
Here all the variables are considered in the models estimated on the
training samples.}

\item{sizemod}{a matrix of dimension \code{N} times  \code{length(methods)}.
Each column gives the number of variables selected at each replication.}

\item{pvarsel}{a matrix of dimension \code{p} times \code{length(methods)}.
 Each column gives the proportion of selection of the input variables among
  the \code{N} replications.}
}
\description{
This function performs the mean squared predicted error (MSE) of parametric,
semi-parametric or non parametric regression models (including possibly
variables selection)  using a repeated learning/test samples approach.
The models are estimated with different methods (chosen by the user) for
comparison purpose. The following methods (with and without variables
selection) are available: multiple linear regression (\code{linreg}),
sliced inverse regression associated with kernel regression (\code{sir}),
random forests (\code{rf}), principal components regression  (\code{pcr}),
partial least squares regression (\code{plsr}), ridge regression (\code{ridge}).
The procedure for variables selection is the same for all the
estimation methods. The importance of each variable is measured  by
estimating the response variable with some perturbations of the covariates
and computing the error due to these perturbations. Then the variables
are selected automatically using an approach  to detect a single change
point position (in mean and variance) in the ordered sequence of VI’s values.
 It is also possible to choose the number of selected variables.
}
\details{
The parameters of the methods (\code{sir}, \code{pcr},
 \code{plsr} and \code{ridge}) are tuned
 on the training samples. The bandwidth for Kernel Regression Smoother is
 tuned by leave one out cross validation. The number of components retained
 in the model for \code{pcr} and \code{plsr} is tuned as follows: for
 each possible number of components, the root mean square error (RMSE) is calculated
 vy 5-fold cross validation  and the number of components is selected by detecting
 a change point position (in mean and variance).
The parameter \code{mtry} for random forests is not tuned and fixed to p/3.
 The number of trees is not tuned and fixed to \code{ntree=300}.
}
\examples{
data(simus)
X <- simus$X
Y <- simus$Y
#res <- choicemod(X,Y,method=c("linreg","sir"), N = 50, nperm = 100)
res <- simus$res1
boxplot(res)

}
\seealso{
\code{\link{boxplot.choicemod}}, \code{\link{barplot.choicemod}},
\code{\link{varimportance}}
}
