% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varimp_sir.R
\name{cv_bandwidth}
\alias{cv_bandwidth}
\title{Tune bandwidth}
\usage{
cv_bandwidth(x, y, hmin = (max(x) - min(x))/20, hmax = (max(x) - min(x))/2,
  nbh = 25, graph.CV = TRUE)
}
\arguments{
\item{x}{a vector with the input values}

\item{y}{a vector with the output values}

\item{hmin}{the minimum bandwidth}

\item{hmax}{the maximum bandwidth}

\item{nbh}{size of the grid of bandwidth}

\item{graph.CV}{If TRUE plot the CV MSE against bandwidth values}
}
\description{
\code{barplot} tunes by LOO cross-validation
the bandwidth for Kernel Regression Smoother.
}
