% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.varimportance.R
\name{plot.varimportance}
\alias{plot.varimportance}
\title{Plots of the importance of the variables}
\usage{
\method{plot}{varimportance}(x, choice = "boxplot", cutoff = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{varimportance}.}

\item{choice}{this should be "boxplot" (for parallel boxplots of the importance of the variables), "meanplot" (for the plot of the mean importance of the variables).}

\item{cutoff}{if TRUE, the variables are ordered by decreasing mean importance and a single change point position (in mean and variance) is indicated by a green vertical line.}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
This function provides the parallel boxplots of the importance of the variables (VI) and the plot of the mean importance of the variables. The baseline value of importance is indicated by a red horizontal line for comparison purpose. If \code{cutoff=TRUE} the variables are ordered by decreasing mean importance and a single change point position (in mean and variance) is indicated by a green vertical line.
}
\examples{
data("simus")
X <- simus$X
Y <- simus$Y1
imp <- varimportance(X,Y,method="linreg",nperm=8)
plot(imp,choice="boxplot", col="lightgreen") #by default
plot(imp,choice="meanplot", cutoff = FALSE)
plot(imp,choice="meanplot", cutoff = TRUE)

}
\seealso{
\code{\link{varimportance}}, \code{\link{select.varimportance}}
}
