% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{psbst}
\alias{psbst}
\title{Meat tenderness data}
\format{A list with the following elements:
\describe{
  \item{dat}{a numerical matrix with 71 rows (one row for each young bull)
  and 22 columns (the 21 first columns are biomarkers, the last one
  measures meat tenderness.}
  \item{res}{results of the \code{choicemod} function for this dataset}
  \item{imp1}{results of the \code{varimportance} function with \code{linreg}}
  \item{imp2}{results of the \code{varimportance} function with \code{sir}}
  \item{imp3}{results of the \code{varimportance} function with \code{rf}}
}}
\source{
These experimental data were obtained on animals coming from the EU FP6
Integrated Project ProSafeBeef (FOODCT-2006-36241).
}
\description{
This dataset describes 71 young bulls on 21 muscular biomarkers
and a numerical variable measuring meat tenderness. A short description of the variables is available here :
\url{https://chavent.github.io/modvarsel/modvarsel-intro.html}
}
