% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.varimportance.R
\name{select}
\alias{select}
\title{Variables selection}
\usage{
select(imp, ...)
}
\arguments{
\item{imp}{an object with the importance of variables.}

\item{...}{additional arguments.}
}
\description{
\code{select} is a generic function for selection of covariates
in a prediction model.
}
