% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.varimportance.R
\name{select.varimportance}
\alias{select.varimportance}
\title{Automatic selection of variables}
\usage{
\method{select}{varimportance}(imp, cutoff = TRUE, nbsel = NULL, ...)
}
\arguments{
\item{imp}{an object of class \code{varimportance}.}

\item{cutoff}{if TRUE a change point criterion is used to define a cutoff value and select variables with mean VI above this cutoff.}

\item{nbsel}{if \code{cutoff=FALSE} the number of variables to be selected.}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
an object of class \code{list} with two components.
\item{var}{a vector with the names of the selected variables}
\item{indices}{a vector with the indices of the selected variables}
}
\description{
Selection of the p' < p variables with the highest measure of importance (VI). The number p' is either defined by detecting a single change point position (in mean and variance) in the ordered VI values or by fixing a priori the number p' of selected variables.
}
\examples{
data("simus")
X <- simus$X
Y <- simus$Y1
imp <- varimportance(X,Y,method="linreg",nperm=8)
select(imp, cutoff=TRUE) # by default
select(imp, cutoff=FALSE, nbsel=7)

}
\seealso{
\code{\link{varimportance}}, \code{\link{plot.varimportance}}
}
