% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simus}
\alias{simus}
\title{Simulated datasets}
\format{A list with the following elements:
\describe{
  \item{X}{a numerical matrix with n=200 realisations of the p=15 covariates
  generated independantly with uniform distributions.}
  \item{Y1}{the response vector of size 200 obtained from X with the model (M1)}
  \item{Y2}{the response vector of size 200 obtained from X with the model (M2)}
  \item{Xnew}{a numerical matrix of n=10 new observations (not used to build the response vectors)
  of the p=15 covariates}
  \item{res1}{results of the \code{choicemod} function with \code{Y1} as response variable (M1)}
  \item{res2}{results of the \code{choicemod} function with \code{Y2} as response variable (M2)}
}}
\source{
\url{https://chavent.github.io/modvarsel/modvarsel-intro.html}
}
\description{
Simulated data generated from two fictitious model: a parametric linear model
(M1) and a semiparametric model (M2). The vector of parameters associated with each
covariate (in both models) gives only the five first covariates linked with
the response variable.
}
