% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varimportance.R
\name{varimportance}
\alias{varimportance}
\title{Variables importance}
\usage{
varimportance(X, Y, method = "linreg", nperm = 10)
}
\arguments{
\item{X}{the input numerical matrix of dimension nxp}

\item{Y}{the output numerical vector of size n}

\item{method}{a method chosen among \code{"linreg"}, \code{"sir"},
\code{"rf"}, \code{"pcr"},  \code{"plsr"} or  \code{"ridge"}}

\item{nperm}{the number of ramdom permutations of the values of the input
variables.}
}
\value{
An object with S3 class "varimportance" and with the following components:
\item{mat_imp}{ a matrix of dimension \code{nperm} times \code{p}. Each colum  gives the  VI values of given variable.}

\item{base_imp}{basic variable importance value i.e. the MSE with the non pertubed data.}
}
\description{
This function  computes for a given method (\code{"linreg"}, \code{"sir"},
\code{"rf"},  \code{"pcr"},  \code{"plsr"},  \code{"ridge"}) the  importance
of the input variables. The  importance of a
variable (VI) is calculated by randomly permuting its values and calculating
 the predicted mean square error (MSE) of the model estimated with the
 perturbed data. Several permutations will give a several VI values for the
 same variable and then a more suitable idea its importance. The VI values
  can be compared to the so called basic VI value defined as the MSE of the
   model estimated with the non perturbed data.
}
\details{
The following methods are available: multiple linear regression (\code{linreg}),
sliced inverse regression associated with kernel regression (\code{sir}),
random forests (\code{rf}), principal components regression  (\code{pcr}),
partial least squares regression (\code{plsr}), ridge regression (\code{ridge}).
If necessary the parameters are tuned as explained in \code{\link{choicemod}}.

If the variable has an effect on the response variable, the random
permutation of its values will affect the estimation and its measure
of importance will take a high value.
}
\examples{
data("simus")
X <- simus$X
Y <- simus$Y1
imp <- varimportance(X, Y, method = "linreg", nperm=15)
imp$mat_imp
apply(imp$mat_imp,2,mean)
}
\seealso{
\code{\link{plot.varimportance}}, \code{\link{select.varimportance}}, \code{\link{choicemod}}
}
