% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.choicemod.R
\name{barplot.choicemod}
\alias{barplot.choicemod}
\title{Barplots of choicemode}
\usage{
\method{barplot}{choicemod}(res, method = "linreg", type = "varsel",
  xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{res}{an object of class \code{choicemod}.}

\item{method}{a regression methods in \code{res$method} ( \code{"linreg"}, \code{"sir"}, \code{"rf"}, \code{"pcr"},
\code{"plsr"}, \code{"ridge"}).}

\item{type}{if \code{"varsel"} the occurences (in percent)
of selection of each covariates are plotted.
If \code{"modsize"} the occurences (in percent) of
the sizes of the reduced models are plotted.}

\item{xlab}{X axis label  ("size of the reduced model" by default
when when \code{type="modsize"}).}

\item{ylab}{Y axis label ("percent" by default).}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
This function gives two types of barplots for a given regression method: the barplot of the occurences (in percent)
of selection of each covariates  and the barplot
of the occurences (in percent) of the number of selected covariates (size of the reduced model).
}
\examples{
data("simus")
res <- simus$res1
class(res)
res$method
barplot(res, method="sir", type="varsel", las = 2,  main = "sir", col = rgb(1,1,0,0.2))
barplot(res, method="sir", type="sizemod", las = 2, main = "sir", col = rgb(0,1,1,0.2))

}
\seealso{
\code{\link{choicemod}}, \code{\link{boxplot.choicemod}}
}
