% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.choicemod.R
\name{boxplot.choicemod}
\alias{boxplot.choicemod}
\title{Parrallel boxplots of choicemod}
\usage{
\method{boxplot}{choicemod}(res, method = NULL, type = "reduced",
  ylab = NULL, ...)
}
\arguments{
\item{res}{an object of class \code{choicemod}.}

\item{method}{if NULL, all the methods in \code{res$method} are plotted.
Otherwise a subset of these methods (\code{"linreg"}, \code{"sir"}, \code{"rf"}, \code{"pcr"},
\code{"plsr"}, \code{"ridge"}).}

\item{type}{this should be \code{"reduced"} (boxplot for reduced estimated models),
\code{"complete"} (boxplot for complete estimated models) or \code{"both"}.
 By default,  \code{type="reduced"}.}

\item{ylab}{Y axis label.}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
The parallel boxplots are built with the MSE values obtained at each replication
of the procedure implemented in \link{choicemod}. At each replication, each regression
method estimates both the reduced model (with covariates selection)
and the complete models (without covariates selection). The boxplots represent the
test mean square errors (performed on test samples) of each regression method including (or not) covariates selection.
}
\examples{
data("simus")
res <- simus$res1
class(res)
res$method
boxplot(res, type="reduced", col="gold", main="Models with variables selection, N=10")
boxplot(res,type="complete",col=rgb(0,0,1,0.2), main="Models with all the variables, N=10")
boxplot(res,type="both", col = "aquamarine", las = 2, main = "N = 10 replications")

}
\seealso{
\code{\link{choicemod}}, \code{\link{barplot.choicemod}}
}
