% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.varimportance.R
\name{select.varimportance}
\alias{select.varimportance}
\title{Automatic selection of variables}
\usage{
\method{select}{varimportance}(imp, cutoff = TRUE, nbsel = NULL, ...)
}
\arguments{
\item{imp}{an object of class \code{varimportance}.}

\item{cutoff}{if TRUE covariates above the changepoint cutoff value are selected.}

\item{nbsel}{the number of covariates to select (only if \code{cutoff=FALSE}).}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
an object of class \code{list} with two components.
\item{var}{a vector with the names of the selected variables}
\item{indices}{a vector with the indices of the selected variables}
}
\description{
Chooses automaticaly the covariates relevant to predict the response variable.
By default a change point criterion is used to define a cutoff value
 and the covariates with mean VI above this cutoff value are selected.
It is also possible to choose the number of covariates to select rather than
using the changepoint cutoff value.
}
\examples{
data("simus")
X <- simus$X
Y <- simus$Y1
imp <- varimportance(X,Y,method="linreg",nperm=8)
select(imp, cutoff=TRUE) # by default
select(imp, cutoff=FALSE, nbsel=7)

}
\seealso{
\code{\link{varimportance}}, \code{\link{plot.varimportance}}
}
