% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varimportance.R
\name{varimportance}
\alias{varimportance}
\title{Variables importance}
\usage{
varimportance(X, Y, method = "linreg", nperm = 10)
}
\arguments{
\item{X}{a numerical matrix containing the \code{p} variables in the model.}

\item{Y}{a numerical response vector.}

\item{method}{a regression method
(\code{"linreg"}, \code{"sir"}, \code{"rf"}, \code{"pcr"}, \code{"plsr"}, \code{"ridge"}).}

\item{nperm}{the number of random perturbations to perform the importance of the covariates (VI).}
}
\value{
An object with S3 class "varimportance" and the following components:
\item{mat_imp}{a matrix of dimension \code{nperm} times \code{p}
with the VI values for each permuation (in rows) and each covariate (un column).}
\item{base_imp}{the original mean square error.}
}
\description{
This function  computes for a given regression method
(\code{"linreg"}, \code{"pcr"},  \code{"plsr"},  \code{"ridge"},\code{"sir"},
\code{"rf"},  ) the  importance
of the covariates by estimating the response variable with some perturbations
of the covariates and computing the error due to these perturbations.
The  variable importance (VI) of a covariate is  the mean square error (MSE) when
 the values of this variable are randomly permuted. Covariates with higher VI are then more
 important to predict the response variable. This procedure is replicated several times
 giving several values of VI for each covariate.  The original MSE (with no perturbation of the data)
 is also performed for comparison purpose.
}
\examples{
data("simus")
X <- simus$X
Y <- simus$Y1
imp <- varimportance(X, Y, method = "linreg", nperm=15)
plot(imp)
imp$mat_imp
apply(imp$mat_imp,2,mean)
}
\seealso{
\code{\link{plot.varimportance}}, \code{\link{select.varimportance}}, \code{\link{choicemod}}
}
